pyproject.toml
[project]
name = "advanced-blog-writing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "openai>=1.98.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tavily-python>=0.7.10",
]
